Adding class TextColor instead of using macros
Before, when displaying the forth dictionary in color on the terminal,
macros where used. Now use instead a class with inheritance for
displaying or not color (dynamicly instead of recompiling). While
working well for the standalone forth, this class, unfortunatly, is not
working on gtkview because gtk only inserts tags with text while my
class just insert tags in the flow of strings. In addition cout is
already redirected in gtkview.