Fix fixme/todo/tbd/bugs
Remove useless ; on POCO_DECLARE_EXCEPTION and POCO_IMPLEMENT_EXCEPTION.

Forth:
-- Use enum with a namespace instead of macros: for stack id and
interpreter state.
-- Save and restore the base when including a file and add an unit test.
-- Forth: fix the idea than BASE can choose alone between set and
getting the display base is erroneous: calling BASE inside a script will
get the stack not empty when trying to get the display base so it'll
always set the base and producing a stack underflow.
-- Fix bug in abort(string) erasing the name of the faulty stream.
-- Fix compil error for standalone forth.

Map:
-- Use inherited SimTaDyn exceptions like done with other SimTaDyn
classes.