Fix TextEditor save/close management
When clicking on the exit button of the SimTaDyn main window with
unsaved Forth documents.

Before this patch: the SimTaDyn main window was hidden before a popup
dialog ask the user to save documents but this dialog could be hidden by
other windows.

After the patch: The popup is called before the main window disapears
so it's now unmaskable by other windows OS.

Note: now all FileChoosedialog will start looking for file at the
SimTaDyn directory.

-- ForthEditor: fix missing menu callback 'open a Forth doc'.
-- SimTaDynWindow: catch now the on delete signal to save unsaved
documents before allowing or not the window destruction.
-- TextEditor: CloseLabel now use pointer on the TextEditor rather on
its notebook. Less limitation.
-- TextEditor: dialogSave gets a new param for knowing if user saves
documents when quitting the application or not.