Reworking GraphMemory container class
-- Rename GraphMemory.hpp and split and move it into several
classes/files. Use a generic name instead. tpp for template file and ipp
for implementation file.
-- Simplify the class and splits it into two classes:
1/ the Set class which is similar to std::vector class and its
push_back() method whith the difference that when the container is full
and a new element is added a new container with the double size is not
created and old container copied into the new one but just a piece of
memory is added.
2/ the Collection class which uses the allocation idea of the class Set
with differencet that elements are not necessary contiguous.