Updating the MVC pattern betwen geographic maps, OpenGL renderer and map editor
-- OpenGL.[ch]pp, GLObject.hpp: change namespace and add some OpenGL
context routines in the aim to fix the following bug by doing the
following actions: 1/ Use a gtkmm version < 3.22 which failes creating
an OpenGL context; 2/ Load a map. This produces a segfault because
calling OpenGL code while the context is not present. Now a security
flag avoid calling GLObject when context is not present.
-- Renderer/able.[ch]pp: remove the test openGL example and now a
renderer can draw a renderable and a renderable can be drawn by a
renderer.
-- DrawingArea.[ch]pp: fix indent. Move a routine into OpenGL.[ch]pp.
renderer/able are linked.
-- SimTaDynMap.hpp: update inheritance for a better integration of Model
for the design pattern Model-View-Controller (MVC) by embedding an
abstract listener for notifying View and Controllers of changes.
-- MapEditor.[ch]pp: Cleaning code. Now inherit from Gtk::HBox to be
included easily inside other widgets. Updating the code to a MVC pattern
(WIP) by implementing observers (from a SimTaDyn map).
-- SimTaDynWindow.cpp: adapt the gtk widgets inclusions.
-- Shapefile.[ch]pp: fix concatening filename, simplifying code.