Lazy allocation of continuous blocks for containers
Before: adding the 1000th element in an empty containers force the
allocation of 1000/X blocks (X is the number of elements by block).
After: allocation is now made on "the fly" so adding the 1000th element
in the container will only make a single block allocation.
Only the reserve() fonction force the allocations of all blocks.